pyproject.toml
[project]
name = "langgraph-learning"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.0",
    "httpx>=0.28.1",
    "ipykernel>=6.29.5",
    "langchain>=0.3.26",
    "langchain-google-genai>=2.1.6",
    "langgraph>=0.5.1",
    "langserve>=0.3.1",
    "langsmith>=0.4.4",
    "mcp[cli]>=1.10.1",
    "notebook>=7.4.4",
    "python-dotenv>=1.1.1",
    "requests-html>=0.10.0",
    "yahoo-fin>=0.8.9.1",
    "yfinance>=0.2.65",
]

[tool.uv.workspace]
members = [
    "mcp_stock_prices",
    "stock-agent-mcp",
]
